server/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
neo4j==5.12.0
python-dotenv==1.0.0
pytz==2023.3.post1
typing_extensions==4.8.0
Werkzeug==2.3.7
